pidmod

+++ client/control/test.cpp
@@ -14,7 +14,7 @@ int main(){
 	double vin = 0;
 	double vout = 0;
 	double ref = 5;
-	double period = 50;
+	int period = 50;
 
 	PIDParameters pp(1,2,1,period,ref,0,10);
 	PID pid;
